MODSCHED-16: Fixed JSONb deserialization issue after model change. Refactored code for code reuse.
