+++ src/main/java/org/folio/scheduler/migration/SplitTimersByMethodMigration.java
@@ -3,40 +3,23 @@ package org.folio.scheduler.migration;
 import java.util.HashSet;
 import java.util.List;
 import java.util.UUID;
-import liquibase.change.custom.CustomTaskChange;
 import liquibase.database.Database;
-import liquibase.database.jvm.JdbcConnection;
-import liquibase.exception.ValidationErrors;
-import liquibase.integration.spring.SpringResourceAccessor;
-import liquibase.resource.ResourceAccessor;
 import lombok.extern.log4j.Log4j2;
-import org.apache.commons.lang3.reflect.FieldUtils;
 import org.folio.scheduler.domain.entity.TimerDescriptorEntity;
 import org.folio.scheduler.repository.SchedulerTimerRepository;
 import org.folio.scheduler.service.JobSchedulingService;
-import org.springframework.context.ApplicationContext;
 import org.springframework.transaction.annotation.Transactional;
 
 @Log4j2
-public class SplitTimersByMethodMigration implements CustomTaskChange {
-
-  private ApplicationContext springApplicationContext;
+public class SplitTimersByMethodMigration extends AbstractCustomTaskChangeMigration {
 
   @Override
   @Transactional
   public void execute(Database database) {
-    JdbcConnection connection = (JdbcConnection) database.getConnection();
     var idsOfTimersToSplit = new HashSet<String>();
-    try (var statement = connection.getWrappedConnection().prepareStatement(
-      "SELECT id FROM timer WHERE jsonb_array_length(timer_descriptor -> 'routingEntry' -> 'methods') > 1")) {
-      var resultSet = statement.executeQuery();
-      while (resultSet.next()) {
-        var timerId = resultSet.getString("id");
-        idsOfTimersToSplit.add(timerId);
-      }
-    } catch (Exception e) {
-      throw new RuntimeException("Failed to execute migration " + this.getClass().getSimpleName(), e);
-    }
+    runQuery(database,
+      "SELECT id FROM timer WHERE jsonb_array_length(timer_descriptor -> 'routingEntry' -> 'methods') > 1",
+      resultSet -> idsOfTimersToSplit.add(resultSet.getString("id")));
 
     if (!idsOfTimersToSplit.isEmpty()) {
       log.info("Found {} timers with multiple HTTP methods - splitting", idsOfTimersToSplit.size());
@@ -68,31 +51,5 @@ public class SplitTimersByMethodMigration implements CustomTaskChange {
         }
       }
     }
-  }
-
-  @Override
-  public String getConfirmationMessage() {
-    return "Completed " + this.getClass().getSimpleName();
-  }
-
-  @Override
-  public void setUp() {
-    // Do nothing
-  }
-
-  @Override
-  public void setFileOpener(ResourceAccessor resourceAccessor) {
-    try {
-      var springResourceAccessor = (SpringResourceAccessor) resourceAccessor;
-      springApplicationContext =
-        (ApplicationContext) FieldUtils.readField(springResourceAccessor, "resourceLoader", true);
-    } catch (IllegalAccessException e) {
-      throw new RuntimeException("Failed to obtain Spring Application Context", e);
-    }
-  }
-
-  @Override
-  public ValidationErrors validate(Database database) {
-    return null;
   }
 }
