+++ src/main/java/org/folio/scheduler/migration/UnscheduleDuplicatesMigration.java
@@ -1,32 +1,22 @@
 package org.folio.scheduler.migration;
 
 import java.util.HashSet;
-import liquibase.change.custom.CustomTaskChange;
 import liquibase.database.Database;
-import liquibase.database.jvm.JdbcConnection;
 import liquibase.exception.CustomChangeException;
-import liquibase.exception.SetupException;
-import liquibase.exception.ValidationErrors;
-import liquibase.integration.spring.SpringResourceAccessor;
-import liquibase.resource.ResourceAccessor;
 import lombok.extern.log4j.Log4j2;
-import org.apache.commons.lang3.reflect.FieldUtils;
 import org.quartz.JobKey;
 import org.quartz.Scheduler;
 import org.quartz.SchedulerException;
-import org.springframework.context.ApplicationContext;
 import org.springframework.transaction.annotation.Transactional;
 
 @Log4j2
-public class UnscheduleDuplicatesMigration implements CustomTaskChange {
-
-  private ApplicationContext springApplicationContext;
+public class UnscheduleDuplicatesMigration extends AbstractCustomTaskChangeMigration {
 
   @Override
   @Transactional
   public void execute(Database database) throws CustomChangeException {
-    JdbcConnection connection = (JdbcConnection) database.getConnection();
-    try (var statement = connection.getWrappedConnection().prepareStatement("""
+    var idsOfTimersToUnschedule = new HashSet<String>();
+    runQuery(database, """
       SELECT
         id
       FROM (
@@ -39,54 +29,19 @@ public class UnscheduleDuplicatesMigration implements CustomTaskChange {
       ) AS tmp
       WHERE
         rn > 1
-      """)) {
-      var resultSet = statement.executeQuery();
-      var idsOfTimersToUnschedule = new HashSet<String>();
-      while (resultSet.next()) {
-        String timerId = resultSet.getString("id");
-        idsOfTimersToUnschedule.add(timerId);
-      }
+      """, resultSet -> idsOfTimersToUnschedule.add(resultSet.getString("id")));
 
-      if (!idsOfTimersToUnschedule.isEmpty()) {
-        log.info("Found {} duplicate timers - unscheduling", idsOfTimersToUnschedule.size());
-        var scheduler = springApplicationContext.getBean(Scheduler.class);
-        idsOfTimersToUnschedule.forEach(timerId -> {
-          try {
-            log.info("Unscheduling timer {}", timerId);
-            scheduler.deleteJob(JobKey.jobKey(timerId));
-          } catch (SchedulerException e) {
-            log.error("Failed to unschedule timer {}", timerId, e);
-          }
-        });
-      }
-    } catch (Exception e) {
-      throw new RuntimeException("Failed to execute migration " + this.getClass().getSimpleName(), e);
+    if (!idsOfTimersToUnschedule.isEmpty()) {
+      log.info("Found {} duplicate timers - unscheduling", idsOfTimersToUnschedule.size());
+      var scheduler = springApplicationContext.getBean(Scheduler.class);
+      idsOfTimersToUnschedule.forEach(timerId -> {
+        try {
+          log.info("Unscheduling timer {}", timerId);
+          scheduler.deleteJob(JobKey.jobKey(timerId));
+        } catch (SchedulerException e) {
+          log.error("Failed to unschedule timer {}", timerId, e);
+        }
+      });
     }
-  }
-
-  @Override
-  public String getConfirmationMessage() {
-    return "Completed " + this.getClass().getSimpleName();
-  }
-
-  @Override
-  public void setUp() throws SetupException {
-    // Do nothing
-  }
-
-  @Override
-  public void setFileOpener(ResourceAccessor resourceAccessor) {
-    try {
-      var springResourceAccessor = (SpringResourceAccessor) resourceAccessor;
-      springApplicationContext =
-        (ApplicationContext) FieldUtils.readField(springResourceAccessor, "resourceLoader", true);
-    } catch (IllegalAccessException e) {
-      throw new RuntimeException("Failed to obtain Spring Application Context", e);
-    }
-  }
-
-  @Override
-  public ValidationErrors validate(Database database) {
-    return null;
   }
 }
