+++ src/main/resources/hibernate.properties
@@ -1,0 +1,1 @@
+hibernate.types.jackson.object.mapper=org.folio.scheduler.configuration.HibernateObjectMapperSupplier
